feat: SI difference between 2 regions of a matrix
For a matrix, we select a 2 points to serve as centers for 2 circular regions. We calculate the average SI for each region and find the difference from them.

This is done to the original phantom images and the reconstructed images to allow us to glimpse into how well SI difference is preserved.